Github Install Generator uses .ruby-version
The ruby/setup-ruby@v1 action will default to using
the .ruby-version file, if no explicit ruby version is specified.
This behaviour can also be invoked by exlicitly specifying
'.ruby-version'.

This change updates the generator to use this by default,
resulting in a single source-of-truth for the project ruby version.
If the ruby-version file is missing, it will display a warning to
the user.